Create io,js file.
