+++ generate-certificates.js
@@ -3,53 +3,16 @@
  */
 'use strict';
 
-import * as fs from 'fs/promises';
-import {createReadStream} from 'fs';
-import {finished} from 'stream';
-import {promisify} from 'util';
 import {join} from 'path';
-import csv from 'csv-parse';
-
-const asyncFinished = promisify(finished);
+import {writeJSON, getDir, getCSV} from './io.js';
 
 const conditionsList = 'WHO list of vaccinable Conditions.csv';
-
-/**
- * Gets the WHO Data dir used to produce certificates.
- *
- * @param {string} path - A path to the who data dir.
- *
- * @throws {Error} - Throws if the dir is not found or is empty.
- *
- * @returns {Promise<Array<string>>} Dir and file names.
-*/
-async function getDir(path) {
-  const directory = await fs.readdir(path);
-  if(directory.length <= 0) {
-    throw new Error(`Dir ${path} is empty`);
-  }
-  return directory;
-}
 
 function getVaccines({records, start = 1}) {
   // the sections of the csv are separated by empty rows
   const firstEmpty = records.findIndex(r => r.every(e => e === ''));
   // so we return all the vaccines before the first empty row
   return records.slice(start, firstEmpty);
-}
-
-async function getCSV({path, parser = new csv.Parser()}) {
-  const records = [];
-  const fileStream = createReadStream(path).pipe(parser);
-  fileStream.on('readable', function() {
-    let record = fileStream.read();
-    while(record != null) {
-      records.push(record);
-      record = fileStream.read();
-    }
-  });
-  await asyncFinished(fileStream);
-  return records;
 }
 
 function createVC(vaccine) {
@@ -143,7 +106,7 @@ async function generateCertificates() {
     await Promise.all(vaccineList.map(v => {
       const {fileName, certificate} = createVC(v);
       const filePath = join(paths.certificates, fileName);
-      return fs.writeFile(filePath, JSON.stringify(certificate, null, 2));
+      return writeJSON({path: filePath, data: certificate});
     }));
   } catch(e) {
     console.error(e);
